Fix cc.requests.outstanding.gauge when using puma web server
https://github.com/cloudfoundry/cloud_controller_ng/issues/1312 introduced cc.requests.outstanding.gauge
which holds the counter in memory. With the introduction of puma there may be multiple processes, each
would have its own value for this metric. This would cause the gauge to flop between values.

This fix, specifically for puma instead reads the value from prometheus and emits it.

Additionally, add a missing mutux to counter for thread safety